read yml files; adding tests
