+++ package.json
@@ -2,16 +2,16 @@
   "name": "team-sync",
   "version": "0.1.0",
   "private": true,
-  "description": "GitHub action to sync ynchronize GitHub Teams with the contents of a teams document",
+  "description": "GitHub action to synchronize GitHub Teams with the contents of a teams document",
   "main": "lib/main.js",
   "scripts": {
     "build": "tsc",
     "format": "prettier --write **/*.ts",
     "format-check": "prettier --check **/*.ts",
     "lint": "eslint src/**/*.ts",
-    "pack": "ncc build",
+    "package": "ncc build --source-map --license licenses.txt",
     "test": "jest",
-    "all": "npm run build && npm run format && npm run lint && npm run pack && npm test"
+    "all": "npm run build && npm run format && npm run lint && npm run package && npm test"
   },
   "repository": {
     "type": "git",
@@ -25,23 +25,22 @@
   "author": "Richard Macklin",
   "license": "MIT",
   "dependencies": {
-    "@actions/core": "^1.2.0",
-    "@actions/github": "^2.1.1",
-    "@sindresorhus/slugify": "^0.11.0"
+    "@actions/core": "^1.4.0",
+    "@actions/github": "^5.0.0",
+    "@sindresorhus/slugify": "^0.11.0",
+    "js-yaml": "^4.1.0"
   },
   "devDependencies": {
-    "@types/jest": "^24.0.23",
-    "@types/node": "^12.7.12",
-    "@typescript-eslint/parser": "^2.8.0",
-    "@zeit/ncc": "^0.20.5",
-    "eslint": "^5.16.0",
-    "eslint-plugin-github": "^2.0.0",
-    "eslint-plugin-jest": "^22.21.0",
-    "jest": "^24.9.0",
-    "jest-circus": "^24.9.0",
-    "js-yaml": "^3.13.1",
-    "prettier": "^1.19.1",
-    "ts-jest": "^24.2.0",
-    "typescript": "^3.6.4"
+    "@types/js-yaml": "^4.0.2",
+    "@types/node": "^16.3.3",
+    "@typescript-eslint/parser": "^4.28.3",
+    "@vercel/ncc": "^0.29.0",
+    "eslint": "^7.31.0",
+    "eslint-plugin-github": "^4.1.1",
+    "eslint-plugin-jest": "^24.3.6",
+    "jest": "^27.0.6",
+    "prettier": "2.2.1",
+    "ts-jest": "^27.0.3",
+    "typescript": "^4.1.3"
   }
 }
